+++ resources/js/components/ActivityLogFilter.vue
@@ -70,24 +70,11 @@
                 </div>
             </div>
             <div class="form-group row">
-                <label class="col-form-label col-md-3" for="filter-activity-type">
-                    {{ $t('global.element_class') }}:
+                <label class="col-form-label col-md-3" for="filter-activity-rawdata">
+                    {{ $t('main.activity.search_in_raw_data') }}:
                 </label>
                 <div class="col-md-9">
-                    <multiselect
-                        v-model="filters.model"
-                        label="label"
-                        track-by="id"
-                        id="filter-activity-type"
-                        :allowEmpty="true"
-                        :closeOnSelect="false"
-                        :hideSelected="true"
-                        :multiple="true"
-                        :options="allSupportedModels"
-                        :placeholder="$t('global.select.placehoder')"
-                        :select-label="$t('global.select.select')"
-                        :deselect-label="$t('global.select.deselect')">
-                    </multiselect>
+                    <input type="text" class="form-control" id="filter-activity-rawdata" v-model="filters.data_text" />
                 </div>
             </div>
             <button type="submit" class="btn btn-outline-primary">
@@ -97,7 +84,9 @@
         <div class="col-md-4" :class="{'mt-3': !hideFilterPanel}">
             <div class="custom-control custom-switch">
                 <input type="checkbox" class="custom-control-input" id="hide-filter-panel-switch" v-model="hideFilterPanel">
-                <label class="custom-control-label" for="hide-filter-panel-switch">Show/Hide filters</label>
+                <label class="custom-control-label" for="hide-filter-panel-switch">
+                    {{ $t('main.activity.hide_filter_panel') }}
+                </label>
             </div>
         </div>
     </div>
@@ -111,9 +100,13 @@
                 type: Boolean,
                 default: false
             },
+            hiddenPanel: {
+                required: false,
+                type: Boolean,
+                default: false
+            },
         },
         mounted() {
-            this.getModels();
         },
         methods: {
             filterActivity(filters) {
@@ -121,26 +114,17 @@
                     filters: filters
                 });
             },
-            getModels() {
-                $httpQueue.add(() => {
-                    $http.get(`activity/loggable`).then(response => {
-                        this.allSupportedModels.length = 0;
-                        this.allSupportedModels = response.data;
-                    });
-                })
-            }
         },
         data() {
             return {
-                hideFilterPanel: false,
+                hideFilterPanel: this.hiddenPanel,
                 types: ['created', 'updated', 'deleted'],
                 allUsers: this.hideUser ? [] : this.$getUsers(),
-                allSupportedModels: [],
                 filters: {
                     users: [],
                     from: null,
                     to: null,
-                    model: null
+                    data_text: '',
                 }
             }
         },